feat: Implement private routes
The private routes will only be registered to the Router when the user logs in.
When the user tries to access a private route before logging in, the Router will
find no matches and delegate to <Route path="*" >.

One important caveat that I discovered is that you can import from *any*
ReactContext, even if the importing component is not inside the imported
Context's render tree; this will result in all the imported Context's values
becoming undefined.
The worst part about this is that React does not give a single warning/error to
alert you.